Re-organization
Re-organize files and importation.
Started to remove redondant code and build global functions.
